chainntnfs: Move pkscript to chainscan package
